Removed unecessary processes. Updated adsirption, desorption and diffusion with the one from simple cubic.
