--- src/processes/diffusion_fcc110.cpp
@@ -1,6 +1,0 @@
-#include "diffusion_fcc110.h"
-
-DiffusionFCC110::DiffusionFCC110()
-{
-
-}
