--- src/processes/diffusion_fcc110.h
@@ -1,11 +1,0 @@
-#ifndef DIFFUSIONFCC110_H
-#define DIFFUSIONFCC110_H
-
-
-class DiffusionFCC110
-{
-public:
-    DiffusionFCC110();
-};
-
-#endif // DIFFUSIONFCC11_H
