+++ src/properties.h
@@ -39,7 +39,6 @@ public:
 
     inline double getRoughness(){ return m_dRoughness; }
     double classCoverage();
-
     double getMeanDH();
 
 private:
